change to float

+++ rfmix_reader/_fb_reader.c
@@ -11,9 +11,9 @@
 #define MIN(a, b) ((a > b) ? b : a)
 
 // Function to read a chunk of the fb matrix
-void read_fb_chunk(uint8_t *buff, uint64_t nrows, uint64_t ncols,
+void read_fb_chunk(float *buff, uint64_t nrows, uint64_t ncols,
                    uint64_t row_start, uint64_t col_start, uint64_t row_end,
-                   uint64_t col_end, uint8_t *out, uint64_t *strides) {
+                   uint64_t col_end, float *out, uint64_t *strides) {
   uint64_t r, c, ce;
   uint64_t row_size = (ncols + 3) / 4; // in bytes
